requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Babel==2.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.3
django-ipware==5.0.2
django-modeltranslation==0.18.11
django-money==3.4.0
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.2.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
ecdsa==0.18.0
idna==3.4
inflection==0.5.1
install==1.3.5
jwcrypto==1.5.0
Markdown==3.5.1
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.25
Pillow==10.1.0
py-moneyed==3.0
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0
